+++ Sever TCP/TCP Client/TCP Client/Form1.resx
