+++ Sever TCP/TCP Client/TCP Client/Properties/AssemblyInfo.cs
