+++ Sever TCP/TCP Client/TCP Client/Properties/Resources.resx
