+++ Sever TCP/TCP Client/TCP Client/Properties/Settings.settings
