--- TCP Server/TCP Server.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33516.290
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TCP Server", "TCP Server\TCP Server.csproj", "{FF902D0E-7834-4155-B329-AE8A645BB6E4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FF902D0E-7834-4155-B329-AE8A645BB6E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FF902D0E-7834-4155-B329-AE8A645BB6E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FF902D0E-7834-4155-B329-AE8A645BB6E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FF902D0E-7834-4155-B329-AE8A645BB6E4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8CF674C6-0DAD-4BC7-AC46-7BDF884C4F2D}
-	EndGlobalSection
-EndGlobal
